+++ lib/sstls/archive.go
@@ -5,19 +5,25 @@ package sstls
  * Read and Save certs with an archive file
  * By J. Stuart McMurray
  * Created 20240327
- * Last Modified 20240327
+ * Last Modified 20251011
  */
 
 import (
 	"crypto/tls"
 	"crypto/x509"
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
+	"syscall"
 	"time"
 
 	"golang.org/x/tools/txtar"
 )
+
+// ErrCacheFileEmpty indicates the the file passed to LoadCachedCertificate was
+// empty.
+var ErrCacheFileEmpty = errors.New("cache file empty")
 
 // LoadCachedCertificate loads the certificate from the named file, which
 // should have been created with SaveCertificate.
@@ -45,9 +51,7 @@ func LoadCachedCertificate(certFile string) (tls.Certificate, error) {
 
 	/* Try to use it. */
 	if 0 == len(certB) {
-		return tls.Certificate{}, fmt.Errorf(
-			"PEM-encoded certificate missing",
-		)
+		return tls.Certificate{}, ErrCacheFileEmpty
 	} else if 0 == len(keyB) {
 		return tls.Certificate{}, fmt.Errorf(
 			"PEM-encoded key missing",
@@ -78,13 +82,28 @@ func LoadCachedCertificate(certFile string) (tls.Certificate, error) {
 // SaveCertificate saves PEM to the given file.  Directories will be created
 // as needed with 0755 permissions.
 func SaveCertificate(certFile string, certPEM, keyPEM []byte) error {
-	/* Make needed directories. */
-	dn := filepath.Dir(certFile)
-	if err := os.MkdirAll(dn, 0700); nil != err {
-		return fmt.Errorf("making directory %s: %w", dn, err)
+	openFile := func() (*os.File, error) {
+		return os.OpenFile(certFile, os.O_CREATE|os.O_WRONLY, 0600)
 	}
+	/* Try opening the file.  If we don't have enough directories it'll
+	fail and we'll try again. */
+	f, err := openFile()
+	if errors.Is(err, syscall.ENOENT) {
+		/* Don't have all the directories. */
+		dn := filepath.Dir(certFile)
+		if err := os.MkdirAll(dn, 0700); nil != err {
+			return fmt.Errorf("making directory %s: %w", dn, err)
+		}
+		/* Try again. */
+		f, err = openFile()
+	}
+	if nil != err {
+		return fmt.Errorf("opening file: %w", err)
+	}
+	defer f.Close()
+
 	/* Save the cert itself. */
-	if err := os.WriteFile(certFile, txtar.Format(&txtar.Archive{
+	if _, err := f.Write(txtar.Format(&txtar.Archive{
 		Comment: []byte(fmt.Sprintf(
 			"Generated %s",
 			time.Now().Format(time.RFC3339),
@@ -96,12 +115,8 @@ func SaveCertificate(certFile string, certPEM, keyPEM []byte) error {
 			Name: txtarKeyFile,
 			Data: keyPEM,
 		}},
-	}), 0600); nil != err {
-		return fmt.Errorf(
-			"writing to %s: %w",
-			certFile,
-			err,
-		)
+	})); nil != err {
+		return fmt.Errorf("writing certificate: %w", err)
 	}
 
 	return nil